+++ features/keychain/module/keychain.js
@@ -1,4 +1,4 @@
-import ExodusModule from '@exodus/module'
+import KeyIdentifier from '@exodus/key-identifier'
 import { parseDerivationPath } from '@exodus/key-utils'
 import { fromMasterSeed as bip32FromMasterSeed } from '@exodus/bip32'
 import SLIP10 from '@exodus/slip10'
@@ -22,14 +22,12 @@ const MAP_KDF = Object.freeze({
 
 export const MODULE_ID = 'keychain'
 
-export class Keychain extends ExodusModule {
+export class Keychain {
   #masters = Object.create(null)
   #legacyPrivToPub = null
 
   // TODO: remove default param. Use it temporarily for backward compatibility.
-  constructor({ legacyPrivToPub = Object.create(null), logger }) {
-    super({ name: MODULE_ID, logger })
-
+  constructor({ legacyPrivToPub = Object.create(null) }) {
     throwIfInvalidLegacyPrivToPub(legacyPrivToPub)
 
     // Create a safe, cloned, frozen map of what was passed to us.
@@ -74,6 +72,8 @@ export class Keychain extends ExodusModule {
   }
 
   async exportKey({ seedId, keyId, exportPrivate }) {
+    keyId = new KeyIdentifier(keyId)
+
     const hdkey = this.#getPrivateHDKey({ seedId, keyId })
     const privateKey = hdkey.privateKey
     let publicKey = hdkey.publicKey
@@ -147,7 +147,7 @@ const keychainDefinition = {
   id: MODULE_ID,
   type: 'module',
   factory: createKeychain,
-  dependencies: ['legacyPrivToPub', 'logger'],
+  dependencies: ['legacyPrivToPub'],
 }
 
 export default keychainDefinition